+++ src/utils/auth_utils.ts
@@ -9,13 +9,13 @@ import {readYML} from './common';
 import {showConnectRepo} from "./notifications";
 import {createUserWithApi} from "./api_utils";
 import {generateSettings} from "../settings";
-import {reactivateAccountHandler, trackRepoHandler} from "../handlers/commands_handler";
-import {Auth0URLs, contextVariables, ECONNREFUSED, getRepoInSyncMsg, HttpStatusCodes, NOTIFICATION, NOTIFICATION_BUTTON} from "../constants";
+import {trackRepoHandler} from "../handlers/commands_handler";
+import {contextVariables, ECONNREFUSED, getRepoInSyncMsg, HttpStatusCodes, NOTIFICATION, NOTIFICATION_BUTTON} from "../constants";
 import { CodeSyncLogger } from "../logger";
-import { generateAuthUrl } from "./url_utils";
 import { pathUtils } from "./path_utils";
 import { RepoState } from "./repo_state_utils";
 import { UserState } from "./user_utils";
+import { authHandler, reactivateAccountHandler } from "../handlers/user_commands";
 
 
 export const isPortAvailable = async (port: number) => {
@@ -29,11 +29,6 @@ export const isPortAvailable = async (port: number) => {
         });
 };
 
-export const redirectToBrowser = (skipAskConnect=false) => {
-    (global as any).skipAskConnect = skipAskConnect;
-    const authorizeUrl = generateAuthUrl(Auth0URLs.AUTHORIZE);
-    vscode.env.openExternal(vscode.Uri.parse(authorizeUrl));
-};
 
 const postSuccessLoginAddUser = (userEmail: string, accessToken: string) => {
     if (!userEmail) return;
@@ -102,16 +97,16 @@ const checkDeactivatedAccount = (email: string, accessToken: string, statusCode:
 export const isAccountActive = async (email: string, accessToken: string) => {
     const isValidAccount = true;
     const userState = new UserState();
-    userState.set(isValidAccount, false);
+    userState.setValidAccount();
     const userResponse = await createUserWithApi(accessToken);
     if (!userResponse.error) return isValidAccount;
     if (userResponse.statusCode === HttpStatusCodes.UNAUTHORIZED) {
-        userState.set(false, false);
+        userState.setInvalidAccount();
         askAndTriggerSignUp();
         return false;
     }
     const isDeactivated = checkDeactivatedAccount(email, accessToken, userResponse.statusCode);
-	userState.set(isValidAccount, isDeactivated);
+	userState.setDeactivated(isDeactivated);
     if (isDeactivated) return false;
     return !userResponse.error.toString().includes(ECONNREFUSED);
 };
@@ -126,7 +121,7 @@ export const createUser = async (accessToken: string, idToken: string) => {
                 isDeactivated: true
         };
         vscode.window.showErrorMessage(NOTIFICATION.SIGNUP_FAILED);
-        CodeSyncLogger.critical("Error creaing user from API", userResponse.error);    
+        CodeSyncLogger.critical("Error creating user from API", userResponse.error);    
         return {
             success: false,
             isDeactivated: false
@@ -140,12 +135,6 @@ export const createUser = async (accessToken: string, idToken: string) => {
     };
 };
 
-export const logout = () => {
-    const logoutUrl = generateAuthUrl(Auth0URLs.LOGOUT);
-    vscode.env.openExternal(vscode.Uri.parse(logoutUrl));
-    markUsersInactive();
-    return logoutUrl;
-};
 
 export const markUsersInactive = (notify=true) => {
     vscode.commands.executeCommand('setContext', contextVariables.showLogIn, true);
@@ -157,12 +146,11 @@ export const markUsersInactive = (notify=true) => {
     });
     fs.writeFileSync(settings.USER_PATH, yaml.dump(users));
     const userState = new UserState();
-	userState.set(false, false);
+	userState.setInvalidAccount();
     if (!notify) return;
-    setTimeout(() => {
-        vscode.window.showInformationMessage(NOTIFICATION.LOGGED_OUT_SUCCESSFULLY);
-    }, 1000);
+    vscode.window.showInformationMessage(NOTIFICATION.LOGGED_OUT_SUCCESSFULLY);
 };
+
 
 export const askAndTriggerSignUp = () => {
     // Trigger sign up process
@@ -172,11 +160,22 @@ export const askAndTriggerSignUp = () => {
         NOTIFICATION.IGNORE
     ]).then(selection => {
         if (selection === NOTIFICATION.LOGIN) {
-            redirectToBrowser(true);
+            authHandler(true);
         }
     });
 };
 
 const parseJwt = (token: string) => {
     return JSON.parse(Buffer.from(token.split('.')[1], 'base64').toString());
+};
+
+export const postSuccessLogout = () => {
+    markUsersInactive();
+    if ((global as any).socketConnection) {
+        (global as any).socketConnection.close();
+        (global as any).socketConnection = null;
+    }
+    if ((global as any).websocketClient) {
+        (global as any).websocketClient = null;
+    }
 };